requirements.txt
streamlit
langchain
langchain-community
chromadb
python-dotenv
SQLAlchemy
langchain-text-splitters
google-generativeai
langchain-google-genai
langchain-groq
groq